+++ changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 2020-0x-xx • `full history <https://github.com/gorakhargosh/watchdog/compare/v0.10.2...master>`__
 
+- Ensure ObservedWatch.path is a string (`#655 <https://github.com/gorakhargosh/watchdog/pull/651>`_)
 - [inotify] Allow to monitor single file (`#655 <https://github.com/gorakhargosh/watchdog/pull/655>`__)
 - Thanks to our beloved contributors: @brant-ruan